requirements.txt
yfinance
python-telegram-bot
python-environ
schedule
